Major UI Overhaul
Yomi

+++ OOP - Task 1C/Game.cpp
@@ -1,6 +1,6 @@
 #include "Game.h"
 
-Game::Game(const std::string& name, const std::string& desc, int cost, int rating, int ID)
+Game::Game(const std::string& name, const std::string& desc, float cost, int rating, int ID)
 	: name(name), description(desc), cost(cost), ageRating(rating), gameID(ID), likes(0.0), dislikes(0.0)
 {
 }
@@ -19,7 +19,7 @@ const std::string& Game::GetDescription() const
 	return description;
 }
 
-int Game::GetCost() const
+float Game::GetCost() const
 {
 	return cost;
 }
