+++ OOP - Task 1C/Player.h
@@ -8,19 +8,19 @@
 class Player : public User
 {
 	public:
-		Player(const std::string&, const std::string&, const Date&, const int& credits);
+		Player(const std::string&, const std::string&, const Date&, const float& credits);
 		~Player();
 		std::ostream& Write(std::ostream& os);
 		std::vector<LibraryItem*> GetLibrary() const;
-		const int getCredits();
-		void deductCredits(int amt);
-		void addCredits(int amt);
+		const float getCredits();
+		void deductCredits(float amt);
+		void addCredits(float amt);
 		void AddToLibrary(LibraryItem*);
 		void RemoveFromLibrary(LibraryItem*);		
 	private:
 		
 	protected:
 		std::vector<LibraryItem*> library;
-		int credits;
+		float credits;
 };
 std::ostream& operator<<(std::ostream& os, Player& p);
